Remove boilerplate, refactor tests
