--- core/tests/harness/mod.rs
@@ -1,61 +1,0 @@
-use eyre::Result;
-use keystone::keystone::Keystone;
-use keystone::keystone_capnp::keystone_config;
-use std::future::Future;
-use tempfile::NamedTempFile;
-
-pub fn test_harness<F: Future<Output = capnp::Result<()>> + 'static>(
-    config: &str,
-    f: impl FnOnce(Keystone) -> F + 'static,
-) -> Result<()> {
-    let mut message = ::capnp::message::Builder::new_default();
-    let mut msg = message.init_root::<keystone_config::Builder>();
-
-    let temp_db = NamedTempFile::new().unwrap().into_temp_path();
-    let temp_log = NamedTempFile::new().unwrap().into_temp_path();
-    let temp_prefix = NamedTempFile::new().unwrap().into_temp_path();
-    let mut source = keystone_util::build_temp_config(&temp_db, &temp_log, &temp_prefix);
-
-    source.push_str(config);
-
-    keystone::config::to_capnp(&source.parse::<toml::Table>()?, msg.reborrow())?;
-
-    let mut instance = Keystone::new(
-        message.get_root_as_reader::<keystone_config::Reader>()?,
-        false,
-    )?;
-
-    // TODO: might be able to replace the runtime catch below with .unhandled_panic(UnhandledPanic::ShutdownRuntime) if gets stabilized
-    let pool = tokio::task::LocalSet::new();
-    let fut = pool.run_until(async move {
-        tokio::task::spawn_local(async move {
-            instance
-                .init(message.get_root_as_reader::<keystone_config::Reader>()?)
-                .await
-                .unwrap();
-
-            f(instance).await?;
-            Ok::<(), capnp::Error>(())
-        })
-        .await
-    });
-
-    let runtime = tokio::runtime::Runtime::new()?;
-    let result = runtime.block_on(fut);
-    runtime.shutdown_timeout(std::time::Duration::from_millis(1));
-    result.unwrap().unwrap();
-
-    Ok(())
-}
-
-pub fn attach_trace() {
-    tracing_subscriber::fmt()
-        .with_max_level(tracing::Level::DEBUG)
-        .with_ansi(true)
-        .with_timer(tracing_subscriber::fmt::time::OffsetTime::new(
-            time::UtcOffset::UTC,
-            time::format_description::well_known::Rfc3339,
-        ))
-        .with_writer(std::io::stderr)
-        .init();
-}
