--- modules/hello-world/hello_world.rs
@@ -1,22 +1,0 @@
-use crate::hello_world_capnp::root;
-
-pub struct HelloWorldImpl {
-    pub greeting: String,
-}
-
-impl root::Server for HelloWorldImpl {
-    async fn say_hello(
-        &self,
-        params: root::SayHelloParams,
-        mut results: root::SayHelloResults,
-    ) -> Result<(), ::capnp::Error> {
-        tracing::debug!("say_hello was called!");
-        let request = params.get()?.get_request()?;
-        let name = request.get_name()?.to_str()?;
-        let greet = self.greeting.as_str();
-        let message = format!("{greet}, {name}!");
-
-        results.get().init_reply().set_message(message[..].into());
-        Ok(())
-    }
-}
