+++ modules/hello-world/main.rs
@@ -1,94 +1,53 @@
 include!(concat!(env!("OUT_DIR"), "/capnproto.rs"));
 
-pub mod hello_world;
-use crate::hello_world_capnp::config;
 use crate::hello_world_capnp::root;
 use capnp::any_pointer::Owned as any_pointer;
-use capnp_macros::capnproto_rpc;
-use capnp_rpc::{rpc_twoparty_capnp, twoparty, RpcSystem};
-use hello_world::HelloWorldImpl;
-use keystone::module_capnp::module_start;
-use std::cell::RefCell;
-use tracing::Level;
 
-pub struct ModuleImpl {
-    bootstrap: RefCell<Option<keystone::keystone_capnp::host::Client<any_pointer>>>,
-    disconnector: RefCell<Option<capnp_rpc::Disconnector<rpc_twoparty_capnp::Side>>>,
+pub struct HelloWorldImpl {
+    pub greeting: String,
 }
 
-#[capnproto_rpc(module_start)]
-impl module_start::Server<config::Owned, root::Owned> for ModuleImpl {
-    async fn start(&self, config: Reader) -> Result<(), ::capnp::Error> {
-        let client: root::Client = capnp_rpc::new_client(HelloWorldImpl {
-            greeting: config.get_greeting()?.to_string()?,
-        });
-        results.get().set_api(client)?;
+impl root::Server for HelloWorldImpl {
+    async fn say_hello(
+        &self,
+        params: root::SayHelloParams,
+        mut results: root::SayHelloResults,
+    ) -> Result<(), ::capnp::Error> {
+        tracing::debug!("say_hello was called!");
+        let request = params.get()?.get_request()?;
+        let name = request.get_name()?.to_str()?;
+        let greet = self.greeting.as_str();
+        let message = format!("{greet}, {name}!");
+
+        results.get().init_reply().set_message(message[..].into());
         Ok(())
     }
-    async fn stop(&self) -> Result<(), ::capnp::Error> {
-        let r = self.disconnector.borrow_mut().take();
-        if let Some(d) = r {
-            d.await?;
-            Ok(())
-        } else {
-            Err(capnp::Error::from_kind(capnp::ErrorKind::Disconnected))
-        }
+}
+
+impl keystone::Module<hello_world_capnp::config::Owned> for HelloWorldImpl {
+    async fn new(
+        config: <hello_world_capnp::config::Owned as capnp::traits::Owned>::Reader<'_>,
+        _: keystone::keystone_capnp::host::Client<any_pointer>,
+    ) -> capnp::Result<Self> {
+        Ok(HelloWorldImpl {
+            greeting: config.get_greeting()?.to_string()?,
+        })
     }
 }
 
 #[tokio::main(flavor = "current_thread")]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let _: Vec<String> = ::std::env::args().collect();
+    keystone::main::<crate::hello_world_capnp::config::Owned, HelloWorldImpl, root::Owned>(
+        async move {
+            //let _: Vec<String> = ::std::env::args().collect();
 
-    #[cfg(feature = "tracing")]
-    tracing_subscriber::fmt()
-        .with_max_level(Level::DEBUG)
-        .with_writer(std::io::stderr)
-        .with_ansi(true)
-        .init();
-
-    tracing::info!("Module starting up...");
-
-    tokio::task::LocalSet::new()
-        .run_until(async move {
-            tokio::task::spawn_local(async move {
-                let mut set: capnp_rpc::CapabilityServerSet<
-                    ModuleImpl,
-                    module_start::Client<config::Owned, root::Owned>,
-                > = capnp_rpc::CapabilityServerSet::new();
-
-                let module_client: module_start::Client<config::Owned, root::Owned> = set
-                    .new_client(ModuleImpl {
-                        bootstrap: None.into(),
-                        disconnector: None.into(),
-                    });
-
-                let reader = tokio::io::stdin();
-                let writer = tokio::io::stdout();
-
-                let network = twoparty::VatNetwork::new(
-                    reader,
-                    writer,
-                    rpc_twoparty_capnp::Side::Server,
-                    Default::default(),
-                );
-                let mut rpc_system =
-                    RpcSystem::new(Box::new(network), Some(module_client.clone().client));
-
-                let server = set.get_local_server_of_resolved(&module_client).unwrap();
-                let borrow = server.as_ref();
-                *borrow.bootstrap.borrow_mut() =
-                    Some(rpc_system.bootstrap(rpc_twoparty_capnp::Side::Client));
-                *borrow.disconnector.borrow_mut() = Some(rpc_system.get_disconnector());
-
-                tracing::debug!("spawned rpc");
-                rpc_system.await.unwrap();
-            })
-            .await
-        })
-        .await?;
-
-    tracing::info!("RPC gracefully terminated");
-
-    Ok(())
+            #[cfg(feature = "tracing")]
+            tracing_subscriber::fmt()
+                .with_max_level(Level::DEBUG)
+                .with_writer(std::io::stderr)
+                .with_ansi(true)
+                .init();
+        },
+    )
+    .await
 }
