+++ modules/indirect-world/Cargo.toml
@@ -40,5 +40,4 @@ keystone-build.workspace = true
 [dev-dependencies]
 eyre.workspace = true
 tempfile.workspace = true
-keystone-util.workspace = true
 async-byte-channel.workspace = true
