+++ modules/indirect-world/indirect_world.rs
@@ -1,4 +1,5 @@
 use crate::indirect_world_capnp::root;
+use capnp::any_pointer::Owned as any_pointer;
 
 pub struct IndirectWorldImpl {
     pub hello_client: hello_world::hello_world_capnp::root::Client,
@@ -20,5 +21,16 @@ impl root::Server for IndirectWorldImpl {
         let msg = hello_response.get()?.get_reply()?.get_message()?;
         results.get().init_reply().set_message(msg);
         Ok(())
+    }
+}
+
+impl keystone::Module<crate::indirect_world_capnp::config::Owned> for IndirectWorldImpl {
+    async fn new(
+        config: <crate::indirect_world_capnp::config::Owned as capnp::traits::Owned>::Reader<'_>,
+        _: keystone::keystone_capnp::host::Client<any_pointer>,
+    ) -> capnp::Result<Self> {
+        Ok(IndirectWorldImpl {
+            hello_client: config.get_hello_world()?,
+        })
     }
 }
