+++ modules/indirect-world/lib.rs
@@ -1,1 +1,3 @@
 include!(concat!(env!("OUT_DIR"), "/capnproto.rs"));
+
+pub mod indirect_world;
