+++ modules/stateful/Cargo.toml
@@ -39,4 +39,4 @@ keystone-build.workspace = true
 [dev-dependencies]
 eyre.workspace = true
 tempfile.workspace = true
-keystone-util.workspace = true
+toml = "0.8"
